The current implementation of FSController in `storage/fs.go` could be improved in the following way:

1. Replace `TryLock` mechanism with [sync.Cond](https://pkg.go.dev/sync#Cond)
2. Using an slice as a FIFO that keeps track of the file descriptors that were **Freed**. Taking into account that the first appended is the oldest one, that will be the first to be **cleaned**. This will remove the necessity of sorting in the clean process.
3. Adding tests to check validity of these changes in order to detect possible deadlocks, and benchmarks tests to see if these changes actually improved previous implementation.

**Benchmarks on previous version**

```
 /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkAcquire$ github.com/xssnick/tonutils-storage/storage

goos: linux
goarch: amd64
pkg: github.com/xssnick/tonutils-storage/storage
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
=== RUN   BenchmarkAcquire
BenchmarkAcquire
BenchmarkAcquire-8             5         271481025 ns/op         8723393 B/op      19535 allocs/op
PASS
ok      github.com/xssnick/tonutils-storage/storage     4.367s
```

**Benchmarks on new implementation**

```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkAcquire$ github.com/xssnick/tonutils-storage/storage

goos: linux
goarch: amd64
pkg: github.com/xssnick/tonutils-storage/storage
cpu: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
=== RUN   BenchmarkAcquire
BenchmarkAcquire
BenchmarkAcquire-8            31          42850579 ns/op          587318 B/op      13247 allocs/op
PASS
ok      github.com/xssnick/tonutils-storage/storage     4.397s
```

**Benchmark overview**

|Version| Iterations | Nanosecond / op | Bytes / op |  Allocations / op |
|----------|--------------|------------------------|---------------|-----------------------|
| Old| 5 | 271481025 ns/op | 8723393 B/op | 19535 allocs/ op |
| New | 32 | 42850579 ns/op | 587318 B/op | 13247 allocs/op |