@xssnick 2024-07-26
Good job! Thank you! 

Will try and merge it soon